changes as per review comments
